+++ Services/Implementations/NavigationService.cs
@@ -1,27 +1,33 @@
 using System;
 using System.Collections.Generic;
+using Microsoft.Extensions.DependencyInjection;
 using NewLab.Services.Interfaces;
-using NewLab.ViewModels.Base;
 
 namespace NewLab.Services.Implementations
 {
     public class NavigationService : INavigationService
     {
-        private readonly Stack<ViewModelBase> _history = new();
-        private ViewModelBase? _currentViewModel;
+        private readonly IServiceProvider _serviceProvider;
+        private readonly Stack<object> _history = new();
+        private object? _currentViewModel;
 
-        public ViewModelBase? CurrentViewModel => _currentViewModel;
+        public NavigationService(IServiceProvider serviceProvider)
+        {
+            _serviceProvider = serviceProvider;
+        }
+
+        public object? CurrentViewModel => _currentViewModel;
 
         public bool CanGoBack => _history.Count > 0;
 
-        public void NavigateTo<TViewModel>() where TViewModel : ViewModelBase
+        public void NavigateTo<TViewModel>() where TViewModel : class
         {
             if (_currentViewModel != null)
             {
                 _history.Push(_currentViewModel);
             }
 
-            _currentViewModel = Activator.CreateInstance<TViewModel>();
+            _currentViewModel = _serviceProvider.GetRequiredService<TViewModel>();
         }
 
         public void GoBack()