+++ Services/Interfaces/INavigationService.cs
@@ -2,8 +2,9 @@ namespace NewLab.Services.Interfaces
 {
     public interface INavigationService
     {
-        void NavigateTo<TViewModel>() where TViewModel : ViewModels.Base.ViewModelBase;
+        void NavigateTo<TViewModel>() where TViewModel : class;
         void GoBack();
         bool CanGoBack { get; }
+        object? CurrentViewModel { get; }
     }
 }