+++ Views/Controls/DashboardContentControl.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace NewLab.Views.Controls
+{
+    public partial class DashboardContentControl : UserControl
+    {
+        public DashboardContentControl()
+        {
+            InitializeComponent();
+        }
+    }
+}