+++ Views/Controls/FunctionPlaceholderControl.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace NewLab.Views.Controls
+{
+    public partial class FunctionPlaceholderControl : UserControl
+    {
+        public FunctionPlaceholderControl()
+        {
+            InitializeComponent();
+        }
+    }
+}