+++ Views/Windows/SetupView.xaml.cs
@@ -13,7 +13,7 @@ namespace NewLab.Views.Windows
             var vm = serviceProvider.GetRequiredService<SetupViewModel>();
             vm.OnSuccess = () =>
             {
-                var mainWindow = new MainWindow();
+                var mainWindow = new MainWindow(serviceProvider);
                 
                 // When MainWindow closes, show LoginView (not SetupView again)
                 mainWindow.Closed += (s, e) =>